Hoist per-session outbound chain scaffolding into BaseChannelGateway
Lark and weixin gateways each duplicated the previous.catch().then()
.finally() chain plus map set/delete dance. Move it into a
scheduleOutboundChain helper on BaseChannelGateway so both onBridgeEvent
overrides just supply their inner async block.

Also converges lark's bridge error log onto formatSafeError, matching
weixin.
